requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
python-dotenv>=1.0.0
rapidfuzz>=3.0.0
azure-cognitiveservices-speech>=1.25.0
pydantic>=2.0.0
groq>=0.4.0
